requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.0
SQLAlchemy~=2.0.40
psycopg2~=2.9.10
psycopg2-binary~=2.9.10
alembic~=1.15.1
python-dotenv~=1.1.0
requests~=2.32.3
pandas~=2.2.3
nltk~=3.9.1
google-genai~=1.8.0


